@@ -9,21 +9,22 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.1",
+    "axios": "^1.4.0",
     "core-js": "^3.6.5",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^10.0.2",
-    "vue": "^2.6.12",
-    "vue-cookies": "^1.7.4"
+    "vue": "^3.2.4",
+    "vue-cookies": "^1.8.3"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^4.5.4",
     "@vue/cli-plugin-eslint": "^4.5.4",
     "@vue/cli-service": "^4.5.4",
     "babel-eslint": "^10.0.3",
-    "eslint": "^7.8.1",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.12"
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.4.0",
+    "sass": "^1.26.5",
+    "sass-loader": "^8.0.2",
+    "vue-template-compiler": "^2.6.12",
+    "webpack": "^4.0.0"
   },
   "eslintConfig": {
     "root": true,
