@@ -1,20 +1,36 @@
 <template>
-    <button type="button" @click="captcha()">点击我</button>
+  <div>
+    <button type="button" @click="show = !show">点击我</button>
+    <Clicaptcha
+        v-if="show" :callback="callback" :src="src"/>
+  </div>
 </template>
 
 <script>
-import Clicaptcha from "@/components/index.js";
+import Clicaptcha from "@/components/Clicaptcha.vue";
 
 export default {
-    methods: {
-        captcha() {
-            Clicaptcha({
-                src: "http://localhost/clicaptcha-server/clicaptcha.php",
-                callback: () => {
-                    alert("yes");
-                }
-            });
-        }
+  components: {
+    Clicaptcha
+  },
+  data() {
+    return {
+      src: "http://127.0.0.1:8000/clicaptcha.php",
+      show: false
     }
+  },
+  methods: {
+    callback() {
+      alert("yes");
+    },
+    captcha() {
+      // Clicaptcha({
+      //   src: "http://localhost/clicaptcha-server/clicaptcha.php",
+      //   callback: () => {
+      //     alert("yes");
+      //   }
+      // });
+    }
+  }
 };
 </script>
