@@ -1,5 +1,7 @@
 module.exports = {
-	css: {
-		extract: false
-	}
+    css: {
+        extract: false
+    },
+    // 关闭eslint校验
+    lintOnSave: false
 }
